% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add}
\alias{add}
\title{add two numbers}
\usage{
add(a, b, really = TRUE)
}
\arguments{
\item{a}{numeric}

\item{b}{numeric}

\item{really}{logical}
}
\value{
a number
}
\description{
add two numbers
}
\examples{
add(2, 3)
}
