% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delpattern.R
\name{delpattern}
\alias{delpattern}
\title{Delete Files with Matching Names}
\usage{
delpattern(directory, pattern, recursive = FALSE)
}
\arguments{
\item{directory}{A directory to process which includes files}

\item{pattern}{A pattern that will be removed}

\item{recursive}{A logical statement if the function will work recursively or not}
}
\value{
A vector of deleted files
}
\description{
Delete Files with Matching Names
}
\examples{
\dontrun{
delete_files_matching_pattern(tempdir(), "test")
delete_files_matching_pattern(tempdir(), "test", recursive = TRUE)
}
}
