% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_beta.r
\name{estimate_beta}
\alias{estimate_beta}
\title{My second function}
\usage{
estimate_beta(y, x)
}
\arguments{
\item{y}{the dependent variable, a vector.}

\item{x}{the independent variable, another vector.}
}
\value{
estimated beta, the correlation between y and x.
}
\description{
My second function
}
\examples{
estimate_beta(c(1,2,3), c(-5,-7,-9))
}
