% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f1.R
\name{f1}
\alias{f1}
\title{Add F1 to Your Text}
\usage{
f1(text = "default")
}
\arguments{
\item{text}{(character) the text to paste with "F1". Default = "default"}
}
\value{
A character vector of length 1.
}
\description{
Add F1 to Your Text
}
\examples{
F1 <- f1(text = "web")
}
