% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Plot of crash locations}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{A numeric value representing the state ID}

\item{year}{A numeric value representing the year}
}
\value{
NULL. An error is raised if an inexistent state number is entered
}
\description{
It creates a map of the desired state with the location of the fatal crashes
occurred in the desired year. Data with non-consistent geographical values are
filtered out. An error is raised if a inexistent state number is entered
}
\examples{
\donttest{
  fars_map_state(13, 2015)
}

}
