% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Data file loader}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A string representing a file name}
}
\value{
A tbl_df object if the input file exists
}
\description{
It creates a data file object of 'tbl_df' type from the file that is given
by the \code{filename} input parameter. An error is raised if the file name
provided does not exist
}
\examples{
fars_read("data/accident_2014.csv.bz2")

}
