% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Multiple data file loader}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{An array of numeric representing the years referring to the
data to be loaded}
}
\value{
A list object containing the data frames referring to years provided
   as input, if the corresponding files exist; NULL otherwise
}
\description{
It creates a list containing the data objects ('tbl_df' type) corresponding
to the years provided by the \code{years} input parameter. The data include
only the \code{MONTH} and \code{YEAR} fields. NULL is returned for the data
table whose year does not exist
}
\note{
Data files to be loaded must be placed in the same directory of this
   R script!
}
\examples{
fars_read_years( c(2013, 2014) )

}
