% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Data summary generator}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{An array of numeric representing the years referring to the
data to be loaded}
}
\value{
A table object (of 'tbl_df' type) containing the data summary
}
\description{
It creates a 'tbl_df' object containing the summary of the counts of fatal
crashes, grouped by month, for any valid year provided in the \code{years}
input parameter. An error is generated if no valid year is entered
}
\examples{
\donttest{
  fars_summarize_years( c(2013, 2018) )
}

}
