\name{foo}
\alias{foo}
\title{Foo}
\description{Devuelve el sucesor.}
\usage{
foo(x)
}
\arguments{
\item{x}{ Un 'numeric vector'a. }
}
\value{
Un 'numeric vector' con el/los sucesor/es de x.
}
\author{Nombre Apellido}
\examples{
foo(1) # devuelve 2
foo(1:3) # devuelve 2:4
}
\keyword{test}