% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{get_inits}
\alias{get_inits}
\title{Function to generate initial values}
\usage{
get_inits(seed, n.chains = 3)
}
\arguments{
\item{seed}{seed value}

\item{n.chains}{number of chains}
}
\description{
Function to generate initial values
}
