% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDeltaLE.R
\name{getDeltaLE}
\alias{getDeltaLE}
\title{Get change in life expectancy}
\usage{
getDeltaLE(gap_country)
}
\arguments{
\item{gap_country}{character The country of interest. See gapminder::gapminder$country}
}
\value{
numeric
}
\description{
Given a country calculate the difference in years of life expectancy in 2007 versus 1952
}
