% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gooddays.R
\name{Gooddays}
\alias{Gooddays}
\title{A function calculate remaining days to a target date from today. The default is 2080-10-05}
\usage{
Gooddays(year = 2080, month = 10, date = 5)
}
\arguments{
\item{year}{target year}

\item{date}{}
}
\value{
Print a time diff.
}
\description{
A function calculate remaining days to a target date from today. The default is 2080-10-05
}
\examples{
Gooddays(year = 2080, month = 10, date = 5)
}
