% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Test function}
\usage{
hello(x)
}
\arguments{
\item{x}{The name of a person}
}
\value{
The output from \code{\link{print}}
}
\description{
Test function
}
\examples{
hello("Xin")
}
