% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{File name generator}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A numeric value representing the year}
}
\value{
A formatted string representing a data file name that can be used
   by \link{fars_read} to load yearly data
}
\description{
It creates a formatted string needed to load the corresponding data file,
provided an appropriate value of the \code{year} input parameter.
}
\note{
Values of the \code{year} parameter which are not of numeric type will
   be accepted but might produce inexistent file names!
}
\examples{
make_filename(2014)

}
