% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emojiprint.R
\name{print_emoji}
\alias{print_emoji}
\alias{print_emoji.data.frame}
\title{THis function displays a message and yellow emoji when run}
\usage{
print_emoji(x, ...)

\method{print_emoji}{data.frame}(x, ...)
}
\arguments{
\item{x}{a dataframe}

\item{...}{arguements passed to \code{\link{print}}}
}
\description{
THis function displays a message and yellow emoji when run
}
\examples{
print_emoji(mtcars)
}
