% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printer.R
\name{printer}
\alias{printer}
\title{Print funct}
\usage{
printer(r, x)
}
\arguments{
\item{r}{second column}

\item{x}{first column}
}
\value{
A tibble
}
\description{
Print funct
}
\examples{
printer(x = rnorm(5), r = rnorm(5))
}
