% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\docType{data}
\name{sample_data}
\alias{sample_data}
\title{Sample data for Analysis}
\format{
A \code{tibble} with 2 columns, wich are:
\describe{
\item{x}{random normal data}
\item{y}{outcome that is normal based on \code{x}}
}
}
\usage{
sample_data
}
\description{
A simple \code{tibble} thath is for simple testing and running examples
}
\keyword{datasets}
