% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{some_function}
\alias{some_function}
\title{A function for doing something}
\usage{
some_function(force = FALSE)
}
\arguments{
\item{force}{If set to \code{TRUE}, \code{data.txt} will be created on the
user's Desktop if their Desktop exists. If this function is used in an
interactive session the user will be asked whether or not \code{data.txt}
should be created. The default value is \code{FALSE}.}
}
\description{
This function takes some action. It also attempts to create a file on your
desktop called \code{data.txt}. If \code{data.txt} cannot be created a
warning is raised.
}

