% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_fun.R
\name{test_fun}
\alias{test_fun}
\title{test_fun}
\usage{
test_fun(x)
}
\arguments{
\item{x}{A integer object}
}
\value{
Integer object, double input \code{test_fun}
}
\description{
First test function
}
\examples{
in <- 3
out <- test_fun(x = in)
}
