% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F1.R
\name{f1}
\alias{f1}
\title{Title}
\usage{
f1(text = "default")
}
\arguments{
\item{test}{(character) the text to paste with "F1". Default="default"}
}
\value{
A character vector of length 1
}
\description{
Title
}
\examples{
f9<-f1 (text = "simple" )
}
