% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read Fatality Analysis Reporting System}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A string of characters indicating the name of 
the file to be loaded}
}
\value{
This function returns a tibble of n observations and
   m columns from the selected file
}
\description{
This is a function that reads data from US National Highway 
Traffic Safety Administration's found in the working directory
}
\details{
an error message if the file does not exist
}
