% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Months and year of data from the US National Highway Traffic 
Safety Administration.}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{The years written as an integer}
}
\value{
This function returns a list containing one or some tibbles of n 
   observations and two variables that are MONTH and year
}
\description{
This function reads data from the US National Highway Traffic 
Safety Administration and returns the variables month and year
}
