% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Fatal injuries suffered in motor vehicle traffic crashes.}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{The years written as an integer}
}
\value{
This function returns a tibble of twelve observations  
   and n variables that are MONTH and YEAR with the summary 
   of the number of fatal injuries per month
}
\description{
This function summarizes fatal injuries sustained in motor 
vehicle traffic accidents by month
}
\details{
Returns NULL if the year is invalid or not found in the data
}
