% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Data names from the US National Highway Traffic Safety 
Administration}
\usage{
make_filename(year)
}
\arguments{
\item{year}{The year written either as an integer}
}
\value{
This function returns a string of characters with the 
   name of the data for the selected year
}
\description{
This function creates the name of the US National Highway 
Traffic Safety Administration's data set for the selected year
}
