% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{printer}
\alias{printer}
\title{Test a printer}
\usage{
printer(x)
}
\arguments{
\item{x}{stest string}
}
\value{
the output of \code{\link{print}}
}
\description{
Test a printer
}
\examples{
printer('abc')
\dontrun{
print('abc')
}
}
